terraform/lambda_functions/post_confirmation/requirements.txt
psycopg2-binary==2.9.7
boto3==1.28.57 